Move Control to a separate component and into content
