+++ cardigan/stories/components/Buttons/Buttons.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta, Story } from '@storybook/addon-docs/blocks';
-import ControlReadme from '@weco/common/views/components/Buttons/Control/README.md';
+import ControlReadme from '@weco/content/components/Control/README.md';
 import * as stories from './Buttons.stories.tsx';
 
 <Meta title={`Components/Buttons`} />
