+++ common/views/components/Iframe/Iframe.tsx
@@ -1,7 +1,7 @@
 import { Component, createRef, ReactElement } from 'react';
 import styled from 'styled-components';
 import PrismicImage from '@weco/common/views/components/PrismicImage/PrismicImage';
-import Control from '@weco/common/views/components/Buttons/Control/Control';
+import Control from '@weco/content/components/Control';
 import Button from '@weco/common/views/components/Buttons';
 import { cross } from '@weco/common/icons';
 import { ImageType } from '@weco/common/model/image';
